pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "copilot-command-ring"
version = "0.1.0"
description = "Host bridge for the Copilot CLI NeoPixel status ring"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
